Renames the project from RcCar to AndroidBlueBot
