+++ README.md
@@ -1,3 +1,6 @@
+**Android BlueBot**
+
+An Android app that controls an Arduino based bot.
 # ![RC Car Logo](https://github.com/fouliex/AndroidRcCar/blob/master/resources/images/RobotCarLogo.jpg) Android RC Car(In Progress)
 ## Wiki Contents
 1. [**Assemble the RC Car**](https://github.com/fouliex/AndroidRcCar/wiki/Assemble-the-RC-Car)
