+++ app/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdkVersion 25
     buildToolsVersion "25.0.2"
     defaultConfig {
-        applicationId "com.car.rc.example.fouliex.rccar"
+        applicationId "com.bluebot"
         minSdkVersion 15
         targetSdkVersion 25
         versionCode 1
